+++ work/noerror3.R
@@ -1,0 +1,5 @@
+library(tidyverse)
+data <- mtcars |>
+  group_by(hp) |>
+  summarize(y = sum(disp))
+data